Step 54: log_prob_of_pair

+++ model.py
@@ -357,7 +357,7 @@ def decode_generated_sequence(ids: list[int] | NDArray, itos: dict[int, str]):
 def log_prob_of_pair(p_matrix: NDArray, current_id: int, next_id: int):
     """Return the log probability of a single (current, next) bigram."""
 
-    return np.log(p_matrix[current_id, next_id])
+    return float(np.log(p_matrix[current_id, next_id]))
 
 # Step 55 - sum_negative_log_probs (not yet solved)
 # TODO: implement
